% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bag_of_words.r
\name{bagOfWords}
\alias{bagOfWords}
\title{Generate matrix with people as rows and words as columns}
\usage{
bagOfWords(x)
}
\arguments{
\item{x}{a corpus object}
}
\value{
x a matrix
}
\description{
Generate matrix with people as rows and words as columns
}

