% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_gephi.r
\name{buildGephiFromCorpus}
\alias{buildGephiFromCorpus}
\title{Build a Gephi file based on a corpus}
\usage{
buildGephiFromCorpus(corpus, trim = 1, FUN_direction = as.directed,
  filename)
}
\arguments{
\item{corpus}{Corpus}

\item{trim}{level to trim the graph paths.}

\item{filename}{}

\item{undirected}{if graph should be directed or undirected}
}
\value{
boolean
}
\description{
Build a Gephi file based on a corpus
}

