% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_gephi.r
\name{buildGephiFromGraph}
\alias{buildGephiFromGraph}
\title{Build a Gephi file based on graph}
\usage{
buildGephiFromGraph(gr, filename, lai = NULL)
}
\arguments{
\item{gr}{Graph}

\item{filename}{}

\item{lai}{Optional data.frame, with "word" and "lai"}
}
\value{
boolean
}
\description{
Build a Gephi file based on graph
}

