% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_lai_degree.r
\name{compareLAIDegree}
\alias{compareLAIDegree}
\title{Returns a data.frame with nodes LAI and degree}
\usage{
compareLAIDegree(x, trim = 1)
}
\arguments{
\item{x}{a data.frame that conforms to a corpus of lexical availability}

\item{trim}{level for graph}
}
\value{
a data.fram with columns (word,lai,degree)
}
\description{
Returns a data.frame with nodes LAI and degree
}

