% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_corpus.r
\name{generateCorpus}
\alias{generateCorpus}
\title{Generate a base object to analyze corpus as
a graph and its lexical availability}
\usage{
generateCorpus(x, lambda = 0.9)
}
\arguments{
\item{x}{a data.frame with the following structure
\describe{
\item{id}{Identifier for the subject}
\item{order}{Order of the answer}
\item{value}{Word for subject \code{id}, order \code{order}}
}}

\item{lambda}{Lambda coefficient for LAI}
}
\value{
a corpus object
}
\description{
Generate a base object to analyze corpus as
a graph and its lexical availability
}

