% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_from_corpus.r
\name{graphFromCorpus}
\alias{graphFromCorpus}
\title{Generate a Graph from igraph, using a data.frame with corpus structure or a corpus object}
\usage{
graphFromCorpus(x)
}
\arguments{
\item{x}{a corpus object}
}
\value{
a directed graph
}
\description{
Generate a Graph from igraph, using a data.frame with corpus structure or a corpus object
}

