% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lai.r
\name{LAI.lexmath}
\alias{LAI.lexmath}
\title{Lexical availability index, LEXMATH version}
\usage{
LAI.lexmath(x, lambda = 0.9)
}
\arguments{
\item{corpus}{object}
}
\value{
data.frame with words and its lexical availability
}
\description{
Lexical availability index, LEXMATH version
}

