% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_path_weight.r
\name{meanPathWeight}
\alias{meanPathWeight}
\title{Calculates mean path weight for every person
In raw}
\usage{
meanPathWeight(x)
}
\arguments{
\item{x}{Corpus}
}
\value{
data.frame with raw MWP for every person
}
\description{
Calculates mean path weight for every person
In raw
}

