% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_statistics.r
\name{nodeStatistics}
\alias{nodeStatistics}
\title{Returns a data.frame with statistics for each node}
\usage{
nodeStatistics(corpus, trim = 1)
}
\arguments{
\item{corpus}{}

\item{trim}{level for graph}
}
\value{
a data.frame with columns (word,lai,degree, weight.sum, weight.mean)
}
\description{
Returns a data.frame with statistics for each node
}

