% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_weight.r
\name{nodeWeight}
\alias{nodeWeight}
\title{Calculates the sum and mean of edge weight for all nodes}
\usage{
nodeWeight(x)
}
\arguments{
\item{x}{graph}
}
\value{
data.frame with nodes and sums on weights
}
\description{
Calculates the sum and mean of edge weight for all nodes
}

