% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/person_lai.r
\name{personalLAI}
\alias{personalLAI}
\title{Calculates mean LAI for every person
In raw and normalized version}
\usage{
personalLAI(x, ...)
}
\arguments{
\item{x}{Corpus}
}
\value{
data.frame with raw and normalized mean LAI for every person
}
\description{
Calculates mean LAI for every person
In raw and normalized version
}

