% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/semanticfields.r
\docType{package}
\name{semanticfields}
\alias{semanticfields}
\alias{semanticfields-package}
\title{Methods to analyze Lexical Availability Index and Word and Person related graphs
based on responses on Lexical Availability exercises}
\description{
Methods to analyze Lexical Availability Index and Word and Person related graphs
based on responses on Lexical Availability exercises
}

