% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sf_wide_csv_to_corpus.r
\name{sfWideCsvToCorpus}
\alias{sfWideCsvToCorpus}
\title{Transform wide representation (each line a set of responses)
on a csv file to corpus object representation}
\usage{
sfWideCsvToCorpus(filename, center_of_interest = "_ci_", other.vars)
}
\arguments{
\item{filename}{CSV filename}

\item{other.vars}{set other variables, as a list or data.frame}

\item{center}{of interest for the corpus}
}
\value{
a corpus object
}
\description{
Transform wide representation (each line a set of responses)
on a csv file to corpus object representation
}

