% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{trimByWeight}
\alias{trimByWeight}
\title{Trim a graph using its arcs weights}
\usage{
trimByWeight(gr, p = 2)
}
\arguments{
\item{gr}{graph}

\item{p}{minimum weight accepted before trim}
}
\description{
Trim a graph using its arcs weights
}

