% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim_by_weight.r
\name{trimByWeight}
\alias{trimByWeight}
\title{Trim by weight}
\usage{
trimByWeight(gr, p = 2)
}
\arguments{
\item{gr}{Corpus}

\item{p}{level of trim}
}
\value{
subgraph
}
\description{
Trim by weight
}

